requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
attrs==20.3.0
awslambdaric==1.0.0
backtrader==1.9.76.123
black==20.8b1
botocore==1.20.3
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cryptography==3.3.1
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
freezegun==1.1.0
future==0.18.2
idna==2.10
iniconfig==1.1.1
jmespath==0.10.0
kiwisolver==1.3.1
loguru==0.5.3
matplotlib==3.2.2
mock==4.0.3
mypy-extensions==0.4.3
mysqlclient==2.0.3
numpy==1.20.0
packaging==20.8
pandas==1.2.0
pathspec==0.8.1
Pillow==8.1.0
platformdirs==2.4.0
pluggy==0.13.1
py==1.10.0
pycparser==2.20
pycrypto==2.6.1
pydantic==1.8.2
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-telegram-bot==13.1
pytz==2020.5
ratelimit==2.2.1
regex==2020.11.13
requests==2.25.1
s3transfer==0.3.4
sentry-sdk==1.5.0
simplejson==3.17.2
six==1.15.0
SQLAlchemy==1.3.22
TA-Lib==0.4.19
tableprint==0.9.1
toml==0.10.2
tomli==1.2.2
tornado==6.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.2
wcwidth==0.2.5
zope.interface==5.4.0
